requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
boto3==1.35.39
botocore==1.35.39
cffi==1.17.1
cryptography==43.0.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
